sorted linked list complete
