--- test/queuemanager/LinkedListNodeTest.java
@@ -1,121 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package queuemanager;
-
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-/**
- *
- * @author Sara
- */
-public class LinkedListNodeTest {
-    
-    public LinkedListNodeTest() {
-    }
-    
-    @BeforeClass
-    public static void setUpClass() {
-        LinkedListNode Test;
-    }
-    
-    @AfterClass
-    public static void tearDownClass() {
-    }
-    
-    @Before
-    public void setUp() {
-    }
-    
-    @After
-    public void tearDown() {
-    }
-//
-//    /**
-//     * Test of setNext method, of class LinkedListNode.
-//     */
-//    @Test
-//    public void testSetNext() {
-//        System.out.println("setNext");
-//        LinkedListNode instance = null;
-//        instance.setNext(null);
-//        // TODO review the generated test code and remove the default call to fail.
-//        fail("The test case is a prototype.");
-//    }
-//
-//    /**
-//     * Test of getItem method, of class LinkedListNode.
-//     */
-//    @Test
-//    public void testGetItem() {
-//        System.out.println("getItem");
-//        LinkedListNode instance = null;
-//        Object expResult = null;
-//        Object result = instance.getItem();
-//        assertEquals(expResult, result);
-//        // TODO review the generated test code and remove the default call to fail.
-//        fail("The test case is a prototype.");
-//    }
-//
-//    /**
-//     * Test of getPriority method, of class LinkedListNode.
-//     */
-//    @Test
-//    public void testGetPriority() {
-//        System.out.println("getPriority");
-//        LinkedListNode instance = null;
-//        int expResult = 0;
-//        int result = instance.getPriority();
-//        assertEquals(expResult, result);
-//        // TODO review the generated test code and remove the default call to fail.
-//        fail("The test case is a prototype.");
-//    }
-//
-//    /**
-//     * Test of setItem method, of class LinkedListNode.
-//     */
-//    @Test
-//    public void testSetItem_GenericType() {
-//        System.out.println("setItem");
-//        Object item = null;
-//        LinkedListNode instance = null;
-//        instance.setItem(item);
-//        // TODO review the generated test code and remove the default call to fail.
-//        fail("The test case is a prototype.");
-//    }
-//
-//    /**
-//     * Test of setItem method, of class LinkedListNode.
-//     */
-//    @Test
-//    public void testSetItem_int() {
-//        System.out.println("setItem");
-//        int priority = 0;
-//        LinkedListNode instance = null;
-//        instance.setItem(priority);
-//        // TODO review the generated test code and remove the default call to fail.
-//        fail("The test case is a prototype.");
-//    }
-//
-//    /**
-//     * Test of getNext method, of class LinkedListNode.
-//     */
-//    @Test
-//    public void testGetNext() {
-//        System.out.println("getNext");
-//        LinkedListNode instance = null;
-//        LinkedListNode expResult = null;
-//        LinkedListNode result = instance.getNext();
-//        assertEquals(expResult, result);
-//        // TODO review the generated test code and remove the default call to fail.
-//        fail("The test case is a prototype.");
-//    }
-    
-}
