+++ test/queuemanager/SortedArrayPriorityQueueTest.java
@@ -5,6 +5,8 @@
  */
 package queuemanager;
 
+import queuemanager.Person;
+import queuemanager.SortedArrayPriorityQueue;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
